:sparkles: fix all view for horizontal mode
