+++ thyrel-web/src/components/GameBar.tsx
@@ -2,6 +2,7 @@ import { Box } from '@material-ui/core';
 import { useSessionContext } from 'hooks/SessionProvider';
 import PlayerCount from './room/PlayerCount';
 import StepTimer from './room/StepTimer';
+import MediaqueryHeight from 'styles/breakpoint';
 
 type GameBarProps = {
   onFinish?: () => void;
@@ -11,14 +12,19 @@ export default function GameBar({ onFinish }: GameBarProps) {
   const { session } = useSessionContext();
   return (
     <Box
-      display="flex"
-      flexWrap="wrap"
-      flexDirection="row"
-      alignItems="center"
-      p={1}
-      px={{ xs: 2, sm: 4 }}
-      justifyContent="space-between"
-      width="100%">
+      sx={{
+        width: 1,
+        justifyContent: 'space-between',
+        px: { xs: 2, sm: 4 },
+        p: 1,
+        alignItems: 'center',
+        flexDirection: 'row',
+        flexWrap: 'wrap',
+        display: 'flex',
+        [MediaqueryHeight.SM]: {
+          position: 'absolute',
+        },
+      }}>
       <Box width={{ xs: 32, sm: 64 }}>
         <PlayerCount
           count={session?.playerFinished || 0}
