+++ thyrel-web/src/pages/room/Draw.tsx
@@ -1,26 +1,45 @@
 import { useMediaQuery } from '@material-ui/core';
-import DrawDesktop from 'components/room/draw/desktop/DrawDesktop';
-import DrawMobileHorizontal from 'components/room/draw/mobile/horizontal/DrawMobileHorizontal';
-import DrawMobileVertical from 'components/room/draw/mobile/vertical/DrawMobileVertical';
+import Loading from 'components/Loading';
 import { useDisableBodyOverflow } from 'components/room/draw/useDisableBodyOverflow';
 import useMobileHorizontal from 'hooks/useMobileHorizontal';
+import { lazy, Suspense } from 'react';
 import theme from 'theme';
+
+const DrawDesktopLazy = lazy(
+  () =>
+    import(
+      /* webpackChunkName: "DrawDesktop" */ 'components/room/draw/desktop/DrawDesktop'
+    ),
+);
+const DrawMobileVerticalLazy = lazy(
+  () =>
+    import(
+      /* webpackChunkName: "DrawMobileVertical" */ 'components/room/draw/mobile/vertical/DrawMobileVertical'
+    ),
+);
+const DrawMobileHorizontalLazy = lazy(
+  () =>
+    import(
+      /* webpackChunkName: "DrawMobileHorizontal" */ 'components/room/draw/mobile/horizontal/DrawMobileHorizontal'
+    ),
+);
 
 export default function Draw() {
   const isDesktop = useMediaQuery(theme.breakpoints.up('md'));
   useDisableBodyOverflow();
 
-  if (isDesktop) {
-    return <DrawDesktop />;
-  }
-  return <DrawMobile />;
+  return (
+    <Suspense fallback={<Loading />}>
+      {isDesktop ? <DrawDesktopLazy /> : <DrawMobile />}
+    </Suspense>
+  );
 }
 
 function DrawMobile() {
   const isHorizontal = useMobileHorizontal();
 
   if (isHorizontal) {
-    return <DrawMobileHorizontal />;
+    return <DrawMobileHorizontalLazy />;
   }
-  return <DrawMobileVertical />;
+  return <DrawMobileVerticalLazy />;
 }
